More lint and tests fixes
